% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{get_t_test}
\alias{get_t_test}
\title{Title of the Function}
\usage{
get_t_test(v_kmer_pos_df, log_geom_values)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of what the function returns
}
\description{
Description of what the function does.
}
